requirements.txt
python-dotenv
openai
openai-agents
google-cloud-aiplatform
google-adk
litellm
requests
beautifulsoup4
